De-duped some state, and a few things started working as expected
