+++ www/UI/App.tsx
@@ -8,18 +8,14 @@ import {
 } from 'react';
 import { KeepAlive } from 'www/KeepAlive';
 import { useJotaiAsyncCallback } from 'www/State/Helpers';
-import { MediaTime, mediaTimeState } from 'www/State/MediaPlaying';
 import { mutedState } from 'www/State/SimpleSavedState';
 import { repeatState, songListState } from 'www/State/SongPlayback';
 import { getStore } from 'www/State/Storage';
+import { MediaTime, mediaTimeState } from 'www/State/TimeState';
 import { MusicDbListener } from 'www/Tools/MusicDbListener';
 import { WebSocketRegistrar } from 'www/Tools/ReactWebSocket';
 import { isValidRefObject } from 'www/Utils';
-import {
-  MaybePlayNext,
-  onClickPlayPause,
-  PlaybackControls,
-} from './PlaybackControls';
+import { MaybePlayNext, PlaybackControls } from './PlaybackControls';
 import { Sidebar } from './Sidebar';
 import { SongPlaying } from './SongPlaying';
 import { ViewSelector } from './Views/Selector';
@@ -32,7 +28,6 @@ type AudioElementProps = { audioRef: React.Ref<HTMLAudioElement> };
 function AudioElement({ audioRef }: AudioElementProps): ReactElement {
   const setMediaTime = useSetAtom(mediaTimeState);
   const isMuted = useAtomValue(mutedState);
-  const onPlayPause = useCallback(() => onClickPlayPause(audioRef), [audioRef]);
   const onEnded = useJotaiAsyncCallback(async (get, set) => {
     /* TODO: What happens when the song ends? */
     // log('Heading to the next song!!!');
@@ -49,36 +44,33 @@ function AudioElement({ audioRef }: AudioElementProps): ReactElement {
     }
   }, []);
 
-  const onTimeUpdate = (ev: SyntheticEvent<HTMLMediaElement>) => {
-    const ae = ev.currentTarget;
-    // log('time update');
-    // log(ev);
-
-    if (!Number.isNaN(ae.duration)) {
-      setMediaTime((prevTime: MediaTime) => {
-        if (
-          Math.trunc(ae.duration) !== Math.trunc(prevTime.duration) ||
-          Math.trunc(ae.currentTime) !== Math.trunc(prevTime.position)
-        ) {
-          return { position: ae.currentTime, duration: ae.duration };
-        } else {
-          return prevTime;
-        }
-      });
-    }
-  };
+  const onTimeUpdate = useCallback(
+    (ev: SyntheticEvent<HTMLMediaElement>) => {
+      const ae = ev.currentTarget;
+      if (!Number.isNaN(ae.duration)) {
+        setMediaTime((prevTime: MediaTime) => {
+          if (
+            Math.trunc(ae.duration) !== Math.trunc(prevTime.duration) ||
+            Math.trunc(ae.currentTime) !== Math.trunc(prevTime.position)
+          ) {
+            return { position: ae.currentTime, duration: ae.duration };
+          } else {
+            return prevTime;
+          }
+        });
+      }
+    },
+    [setMediaTime],
+  );
   return (
     <audio
       // src={songKey !== '' ? '/tune/' + songKey : ''}
       src={'/tune/song.m4a'}
       ref={audioRef}
-      // autoPlay={true}
-      onPlay={onPlayPause}
-      onPause={onPlayPause}
       onEnded={onEnded}
       onTimeUpdate={onTimeUpdate}
       muted={isMuted}
-      controls={true}
+      controls={false}
     />
   );
 }
