+++ www/UI/SongPlaying.tsx
@@ -10,7 +10,7 @@ import {
   useCallback,
   useEffect,
 } from 'react';
-import { useJotaiAsyncCallback, useJotaiCallback } from 'www/State/Helpers';
+import { useJotaiAsyncCallback } from 'www/State/Helpers';
 import {
   mutedState,
   shuffleState,
